rename JoinGameQueue to JoinGameQueueManual
