+++ frontend/src/components/JoinGameQueueManual.vue
